Move exceptions into seperate folder and update use statements
