+++ src/FINDOLOGIC/Export/CSV/CSVExporter.php
@@ -3,7 +3,7 @@
 namespace FINDOLOGIC\Export\CSV;
 
 use FINDOLOGIC\Export\Exporter;
-use FINDOLOGIC\Export\Helpers\BadPropertyKeyException;
+use FINDOLOGIC\Export\Exceptions\BadPropertyKeyException;
 use FINDOLOGIC\Export\Helpers\DataHelper;
 
 class CSVExporter extends Exporter
