+++ src/FINDOLOGIC/Export/Exceptions/PropertyKeyNotAllowedException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace FINDOLOGIC\Export\Exceptions;
+
+class PropertyKeyNotAllowedException extends \RuntimeException
+{
+    public function __construct($key)
+    {
+        $format = 'Property key "%s" is reserved for internal use and overwritten when importing.';
+        parent::__construct(sprintf($format, $key));
+    }
+}
