+++ src/FINDOLOGIC/Export/Exceptions/ValueIsNotIntegerException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace FINDOLOGIC\Export\Exceptions;
+
+class ValueIsNotIntegerException extends \RuntimeException
+{
+    public function __construct($value)
+    {
+        parent::__construct(sprintf('%s is not an integer!', $value));
+    }
+}
