+++ src/FINDOLOGIC/Export/Exceptions/ValueIsNotNumericException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace FINDOLOGIC\Export\Exceptions;
+
+class ValueIsNotNumericException extends \RuntimeException
+{
+    public function __construct()
+    {
+        parent::__construct('Value is not a valid number!');
+    }
+}
