+++ src/FINDOLOGIC/Export/Exceptions/ValueIsNotPositiveIntegerException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace FINDOLOGIC\Export\Exceptions;
+
+class ValueIsNotPositiveIntegerException extends \RuntimeException
+{
+    public function __construct($value)
+    {
+        parent::__construct(sprintf('%s is not an positive integer!', $value));
+    }
+}
