+++ src/FINDOLOGIC/Export/XML/Page.php
@@ -2,16 +2,8 @@
 
 namespace FINDOLOGIC\Export\XML;
 
+use FINDOLOGIC\Export\Exceptions\ItemsExceedCountValueException;
 use FINDOLOGIC\Export\Helpers\XMLHelper;
-
-class ItemsExceedCountValueException extends \RuntimeException
-{
-    public function __construct()
-    {
-        $message = 'The number of items must not exceed the count value';
-        parent::__construct($message);
-    }
-}
 
 class Page
 {
