+++ tests/FINDOLOGIC/Export/Tests/XmlSerializationTest.php
@@ -12,8 +12,8 @@ use FINDOLOGIC\Export\Data\Property;
 use FINDOLOGIC\Export\Data\Url;
 use FINDOLOGIC\Export\Data\Usergroup;
 use FINDOLOGIC\Export\Exporter;
-use FINDOLOGIC\Export\Helpers\InvalidUrlException;
-use FINDOLOGIC\Export\Helpers\UnsupportedValueException;
+use FINDOLOGIC\Export\Exceptions\InvalidUrlException;
+use FINDOLOGIC\Export\Exceptions\UnsupportedValueException;
 use FINDOLOGIC\Export\Helpers\XMLHelper;
 use FINDOLOGIC\Export\XML\Page;
 use FINDOLOGIC\Export\XML\XMLExporter;
@@ -99,7 +99,7 @@ class XmlSerializationTest extends TestCase
     }
 
     /**
-     * @expectedException \FINDOLOGIC\Export\XML\ItemsExceedCountValueException
+     * @expectedException \FINDOLOGIC\Export\Exceptions\ItemsExceedCountValueException
      */
     public function testMoreItemsSuppliedThanCountValueCausesException()
     {
@@ -188,7 +188,7 @@ class XmlSerializationTest extends TestCase
     }
 
     /**
-     * @expectedException \FINDOLOGIC\Export\Data\BaseImageMissingException
+     * @expectedException \FINDOLOGIC\Export\Data\Exceptions\BaseImageMissingException
      */
     public function testMissingBaseImageCausesException()
     {
@@ -203,7 +203,7 @@ class XmlSerializationTest extends TestCase
     }
 
     /**
-     * @expectedException \FINDOLOGIC\Export\Data\ImagesWithoutUsergroupMissingException
+     * @expectedException \FINDOLOGIC\Export\Exceptions\ImagesWithoutUsergroupMissingException
      */
     public function testImagesWithoutUsergroupMissingCausesException()
     {
@@ -410,7 +410,7 @@ class XmlSerializationTest extends TestCase
     }
 
     /**
-     * @expectedException \FINDOLOGIC\Export\Helpers\UnsupportedValueException
+     * @expectedException \FINDOLOGIC\Export\Exceptions\UnsupportedValueException
      * @dataProvider unsupportedValueProvider
      *
      * @param string $method Name of the method to call to interact with an unsupported value.
@@ -439,7 +439,7 @@ class XmlSerializationTest extends TestCase
 
 
     /**
-     * @expectedException \FINDOLOGIC\Export\Helpers\InvalidUrlException
+     * @expectedException InvalidUrlException
      */
     public function testAddingInvalidUrlToImageElementCausesException()
     {
